Update pages files with header and footer
